# openapi.yaml
openapi: 3.0.1
info:
    description: |
        VICOO 个人知识管理助手 API
        
        提供笔记管理、知识图谱、AI问答、标签系统等功能，
        服务于 AI 智能助手的全局知识库问答能力。
    title: VICOO 知识管理助手插件
    version: v1
servers:
    - url: http://localhost:8000
      description: 本地开发服务器
paths:
...
    /api/search:
        get:
            operationId: search_notes
            summary: 搜索笔记
            parameters:
                - name: q
                  in: query
                  required: true
                  schema:
                      type: string
            responses:
                "200":
                    description: 成功
                    content:
                        application/json:
                            schema:
                                type: object
...
    /api/workflow/search:
        post:
            operationId: workflow_search
            summary: 智能搜索工作流
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                query:
                                    type: string
                                platform:
                                    type: string
                                contentType:
                                    type: string
            responses:
                "200":
                    description: 成功
                    content:
                        application/json:
                            schema:
                                type: object
    /api/workflow/parse:
        post:
            operationId: parse_url
            summary: 解析URL内容
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                url:
                                    type: string
            responses:
                "200":
                    description: 成功
                    content:
                        application/json:
                            schema:
                                type: object
    /api/workflow/coze/chat:
        post:
            operationId: coze_chat
            summary: Coze智能体对话
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                query:
                                    type: string
                                conversation_id:
                                    type: string
                                user_id:
                                    type: string
            responses:
                "200":
                    description: 成功
                    content:
                        application/json:
                            schema:
                                type: object

    # ==================== 数据统计 ====================
...
    /api/workflow/coze/config:
        get:
            operationId: get_coze_config
            summary: 获取Coze配置
            responses:
                "200":
                    description: 成功
                    content:
                        application/json:
                            schema:
                                type: object
        post:
            operationId: set_coze_config
            summary: 设置Coze配置
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                apiToken:
                                    type: string
                                botId:
                                    type: string
                                baseUrl:
                                    type: string
            responses:
                "200":
                    description: 成功
                    content:
                        application/json:
                            schema:
                                type: object

    # ==================== 健康检查 ====================
...